Renamed FinishAndStartMain to NavigateToCustomReader
